// js/stat-evaluation-player/src/generated/RushEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A quick possession transition where the attacking team has numbers moving out of its defensive half.
 */
export type RushEvent = { start_time: number, start_frame: number, end_time: number, end_frame: number, is_team_0: boolean, attackers: number, defenders: number, };