// js/stat-evaluation-player/src/generated/RushTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Per-team accumulated rush stats.
 */
export type RushTeamStats = { count: number, two_v_one_count: number, two_v_two_count: number, two_v_three_count: number, three_v_one_count: number, three_v_two_count: number, three_v_three_count: number, };